# dist/swagger_output.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "v1.0.0",
        "title": "Swagger Demo Project",
        "description": "Implementation of Swagger with TypeScript"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": ""
        }
    ],
    "paths": {
        "/api/employee/getEmployees": {
            "get": {
                "description": "",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/employee/getEmployee/{id}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/employee/updateEmployee": {
            "patch": {
                "description": "",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "example": "any"
                                    },
                                    "name": {
                                        "example": "any"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}